Refactor MainScript orchestration
